refactor(app): generic explorer-table-paginated
